requirements.txt
pyside6
pyinstaller
auto-py-to-exe
pillow